hushes the BIL lifter when there is no BIL
This is a quick fix that prevents the BIL lifter from interpreting an
absence of BIL as an absence of effects. The root of the problem is
that we chose `[]` as the empty value for the BIL domain, which is
easy to confuse with an absence of effects (or representable in BIL
effects).

We can't change the domain, as it will break the interface, but a
better solution might be changing the domain of the Bil.code slot that
was introduced recently and is not yet in the stable API. Meantime,
the current solution will be the hotfix.