restores output for non-empty instructions
It looks like that #1281 rendered bap useless for unlifted
instructions. Both mc and objdump commads are now outputing
empty objects even without asm or opcodes, which makes it
really hard to write a new lifter.

This PR fixes it on mc and objdump level. I am not sure if we also
need this in the disassembler driver (as in any case the output of
this driver will be incorrect without having full semantics).