makes KB name to string conversion more robust
The name to string conversion could throw an error when a name is not
interned, which could happen, for example, when the cache written by
one version of the program is read by another version in which some
name is no longer present. In the case of a missing name we represent
it as `id:<repr>` where `<repr>` is the hexadecimal representation of
the missing name hash.