adds armv8 (aarch64) lifter (#1291)
* adds a minimal (incomplete) AARCH64 lifter

* makes KB name to string conversion more robust

The name to string conversion could throw an error when a name is not
interned, which could happen, for example, when the cache written by
one version of the program is read by another version in which some
name is no longer present. In the case of a missing name we represent
it as `id:<repr>` where `<repr>` is the hexadecimal representation of
the missing name hash.

* adds --show-addr and --show-memory options to bap mc and objdump

It makes it much easier to write lifters.

* adds a few more semantic primitives and tweaks the existing

The added semantic primitives are
 - lnot
 - extract
 - concat
 - select

Also tweaked casts and reciprocal to handle static cases. And tweaks
all monoids to choose the most significant (widest) sort instead of
default to the word size.

* lifts even more aarch64 instructions

* updates armv8's register file

* uses Theory.Target.reg to correctly map names to registers

* adds more primitives

* disables handling of pseudo-registers

right now it is not working as expected

* implements more instructions

* fixes function argument passing

* implements variable/place tracking adds a symbol primitive

* adds a Theory.Target.has_roles function

* implements constant and zero registers handling on the theory level

* adds the MIPS ZERO register to the target description

resolves #1176

* adds more aarch64 instructions

* drops any handling of the pseudoregisters from the lisp compiler

* adds more aarch64 instructions

* finishes all the necessary aarch64 instructions

* enables ABI and fixes a few instructions

the trivial executable works, echo is not yet

* more bugfixes - the lifter is now minimally complete

* implements armv8 (aapcs64) ABI

* updates the testsuite