implements the array to pointer conversion in C ABI processor
All arrays are implicitly converted to pointers when passed to a
function, therefore when we see `elt p[]`, `elt p[X]`, or
`elt (*p)[X]` in the prototype we shall convert them to `elt *`.

The conversion is applied in the C ABI processor application function
as it is a feature of the language not of an individual ABI. This will
enable this feature in all C ABI implementations (processors) as they
will not see arrays anymore.