fixes implicit coercion in Primus Lisp monoid operators
Arithmetic and bitwise operators in Primus Lisp accept bitvectors of
arbitrary lengths and performs implicit coercion of the operands to
the greatest sort.

However, the first argument was accidentally removed from the
coercion, which is now fixed. In addition, the coercion procedure was
missing an opportunity to perform the coercion statically, which is
done in the new implementation.