preserves the encoding name of the instruction (#1303)
Ensures that the instruction encoding matches with the name of
encoding that is used to decode it. Before this fix, the encoding name
was sometimes taken from the triple, which enforced us to have the
names of encoding that match with the triple. We can now safely peek
any encoding name.