adds `bap primus-lisp-documentation`
Much like the old `bap --primus-lisp-documentation` this command
prints the Primus Lisp documentation in the .org format. But it
doesn't require a file as it is possible to specify the target using
the `--target` command line option.

Since `primus-lisp-documentation` is a separate command, it is not
necessary to specify a bogus file name and there are more options to
control the output. First of all, it is now possible to specify the
system (using `--system`), so that now it is possible to get the
documentation for symbolic primtives and functions. It is also
possible to set the package name (using `--package`) and request the
documentation for the Primus Lisp Semantics subsystem (with
`--semantics`).

Therefore, to get the documentation for thumb semantics (functions and
primitives), use

```
bap primus-lisp-documentation --semantics --package=thumb
```

The list of packages, along with their documentation is also available
in the `bap primus-lisp-documentation`