adds support for big-endian arm and thumb architectures (#1309)
The big endian thumb and arm targets are [not supported by LLVM][1]
and while [the fix][2] is available for many years it is still not
reviewed. I tried to [raise the attention][3] to it but with no
success. Since it looks like that they are not going to fix it we need
to fix it on our side. The solution is a little bit hacky but we can
keep it until we switch to Ghidra for disassembling/lifting.

Examples,

```
$ bap mc --arch=thumb --show-insn=asm --show-bil -- 1a 42
tst r2, r3
{
  #1 := R2 & R3
  ZF := #1 = 0
  NF := extract:31:31[#1]
}

$ bap mc --arch=thumb --order=big --show-insn=asm --show-bil -- 42 1a
tst r2, r3
{
  #1 := R2 & R3
  ZF := #1 = 0
  NF := extract:31:31[#1]
}

$ file echo
echo: ELF 32-bit MSB executable, ARM, version 1 (ARM), dynamically
linked, interpreter /lib/ld-uClibc.so.0, for GNU/Linux 2.0.0, stripped

$ bap echo -dasm | grep 8c90 -A4
8c90: <main>
8c90:
8c90: e9 2d 47 f0                                 push {r4, r5, r6, r7, r8, r9, r10, lr}
8c94: e5 9f 33 34                                 ldr r3, [pc, #0x334]
8c98: e1 a0 70 01                                 mov r7, r1
8c9c: e1 a0 60 00                                 mov r6, r0
8ca0: e3 a0 00 06                                 mov r0, #6

$ objdump echo -d | grep 8c90 -A4
00008c90 <main@@Base>:
    8c90:	e92d47f0 	push	{r4, r5, r6, r7, r8, r9, sl, lr}
    8c94:	e59f3334 	ldr	r3, [pc, #820]	; 8fd0 <main@@Base+0x340>
    8c98:	e1a07001 	mov	r7, r1
    8c9c:	e1a06000 	mov	r6, r0
    8ca0:	e3a00006 	mov	r0, #6
```

fixes #1299

[1]: https://bugs.llvm.org/show_bug.cgi?id=38721
[2]: https://reviews.llvm.org/D48811
[3]: https://twitter.com/ivg_t/status/1384932479967055877?s=20