removes the old unit tracking module, fixes the modern one
The old unit tracking module, which is used what compilation units are
loaded to the core program, is no longer used since we phased out
compilers that required it. However, the modern unit that relies on
OCaml runtime to track the units was also broken (in all versions of
OCaml since the introduction of this facility in 4.08.0 till the
latest 4.12.0) as it wasn't properly initialized. This PR adds a
kludge that ensures proper initialization.

Note, the bug didn't manifest in BAP as we have several layers of
protection, but it is still better to have it working properly.