switches to FrontC.4.x adds the preprocessing option
Since FrontC.4.x supports new types from C99+ we had to update BAP as
well, both its representation of the C types as well as the parsing
procedure.

We also added an option (disabled by default) to apply
preprocessing (with an ability to specify the path to it) since
FrontC.4.1+ is much more robust and is capable of chewing linux system
headers.