passes full information about the operand type from the backend
In p-code the semantics of an operation is defined by the types of its
operands. In case of the unique variables the type is not known to us
so we have to pass it. This commit extends the previous approach,
where we were passing only the kind of the operand (mem vs imm) to the
full type qualification, where the type of memory is represented with
Nil and the type of immediates is represented by its size in the
number of bits.