improves primitives performance in Primus Lisp
With this optimization Primus Lisp-based lifters run five times
faster. This especially important for ghidra backend lifters, which
are fully dependent on Primus Lisp.

The idea is to let the primitive implementors provide the body of
their primitive so that every primitive is not computed via the
semantics promise but is invoked directly. Another big idea is to
provide such an interface that will allow to factor out computations
that are common to the target. The same idea could be extrapolated to
all promises.