fixes bapdoc (#1345)
Despite that https://github.com/ocaml/odoc/issues/539 is fixed long
time ago it never landed into the 1.x branch of odoc, which is the
stable and the released version. All beta versions do not have the
stack overflow bug on core_kernel, but render unreadable documentation
(tried 2.0.0-beta till the latest master). Therefore, we had to
fallback to the `--no-pkg-deps` workaround, so that we can get at
least the BAP documentation. Note, this workaround means that the
links to core kernel and other external libraries will be broken, only
links across bap universe will work.