fixes a knowledge base race condition in the run plugin (#1347)
The run plugin was resetting the state of the toplevel knowledge base
to the initial and effectively discarding the knowledge that was
accumulated when the Program.t and Symtab.t were constructed as well
discarding any information that might be added to the knowledge base
in project passes.

The run pass was capturing the state at the time when the Project.t is
not fully ready, and at then end was using `Toplevel.set` to reset any
accumualted knowledge.

Effectively, this bug is a race condition, which we could prevent if
the toplevel interface provided safe functions. This will be addressed
in the next couple of pull requests (they will be more aggresive and
change the semantics of several constructs so I decided to split them
for this hotfix).

Big thanks to @anwarmamat for providing a minimal reproducible
example, without it would be nearly impossible to identify it.