fixes variable rewriter and some Primus Lisp symbolic functions (#1350)
* fixes variable rewriter and some Primus Lisp symbolic functions

The variable rewriter was incorrectly setting the mask when a register
was modified through its alias. Instead of preserving the bits that
are not set it was erasing them.

There were also a couple of issues in handling registers and variables
in the Primus Lisp lifter. First of all, let bindings were breaching
the type system by binding variables to values that have different
type. The set function was sometimes mistyping the
registers (reminiscent of times when there were no pseudo-registers).

* allows macos to fail without canceling all other jobs

It fails too often and for random but independent of us reasons.