adds support for namespaces and switches to universal actions
Now an action could be any element oither than `<data>`. The actions
are represented as a pair of a name and an attribute mapping. Both
action names and attribute names are represented with KB.Name.t. The
namespaces are supported but we take only the host part of the XML
namespace name if it is represented as a URI.