adds forward-chaining rules and Primus Lisp methods (#1363)
* adds forward-chaining rules to the knowledge base

* adds some useful primitives for writing primitives to Primus

* evaluate keywords to themselves in Primus Lisp

The keyword symbols, like `:foo` were treated as variables, but we
want them to be symbols. The change is made on the reader (parser)
level so it will affect both Primus Lisp implementations - static
and dynamic.

* allows any primitive to return non-reified symbol

When a primitive returns a value that has the symbol slot set to
something, the interpreter will automatically intern it. It simplifies
writing the primitives as the implementor doesn't need to bother about
interning and setting the static part of the value.

* implements methods in the Primus Lisp semantics interpreter

* updates the testsuite, uses the newer opam syntax